+++ cluster-autoscaler/core/scaledown/actuation/priority_test.go
@@ -183,57 +183,3 @@ func TestGroupByPriority(t *testing.T) {
 	groups := groupByPriority(shutdownGracePeriodByPodPriority, []*apiv1.Pod{p1, p2, p3, p4, p5}, []*apiv1.Pod{p6, p7, p8, p9, p10})
 	assert.Equal(t, wantGroups, groups)
 }
-
-func TestParseShutdownGracePeriodsAndPriorities(t *testing.T) {
-	testCases := []struct {
-		name  string
-		input string
-		want  []kubelet_config.ShutdownGracePeriodByPodPriority
-	}{
-		{
-			name:  "empty input",
-			input: "",
-			want:  nil,
-		},
-		{
-			name:  "Incorrect string - incorrect priority grace period pairs",
-			input: "1:2,34",
-			want:  nil,
-		},
-		{
-			name:  "Incorrect string - trailing ,",
-			input: "1:2, 3:4,",
-			want:  nil,
-		},
-		{
-			name:  "Incorrect string - trailing space",
-			input: "1:2,3:4 ",
-			want:  nil,
-		},
-		{
-			name:  "Non integers - 1",
-			input: "1:2,3:a",
-			want:  nil,
-		},
-		{
-			name:  "Non integers - 2",
-			input: "1:2,3:23.2",
-			want:  nil,
-		},
-		{
-			name:  "parsable input",
-			input: "1:2,3:4",
-			want: []kubelet_config.ShutdownGracePeriodByPodPriority{
-				{1, 2},
-				{3, 4},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			shutdownGracePeriodByPodPriority := ParseShutdownGracePeriodsAndPriorities(tc.input)
-			assert.Equal(t, tc.want, shutdownGracePeriodByPodPriority)
-		})
-	}
-}
